Check the error value rather than a string and rename utils to parseint
