+++ internal/parseint/parse.go
@@ -1,4 +1,4 @@
-package utils
+package parseint
 
 import "strconv"
