+++ x/launch/client/cli/tx.go
@@ -10,7 +10,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/spf13/cobra"
-	"github.com/tendermint/spn/internal/utils"
+	"github.com/tendermint/spn/internal/parseint"
 	"github.com/tendermint/spn/x/launch/types"
 )
 
@@ -111,7 +111,7 @@ func CmdApprove() *cobra.Command {
 			}
 
 			// Convert value for proposal ID
-			proposalID, err := utils.ParseInt32(args[1])
+			proposalID, err := parseint.ParseInt32(args[1])
 			if err != nil {
 				return err
 			}
@@ -147,7 +147,7 @@ func CmdReject() *cobra.Command {
 			}
 
 			// Convert value for proposal ID
-			proposalID, err := utils.ParseInt32(args[1])
+			proposalID, err := parseint.ParseInt32(args[1])
 			if err != nil {
 				return err
 			}
